Refactored code and improved responsiveness, css transition cancellation time and mobile browser experience
Renamed variables and functions to be more readable. The slider and picker now use percentages instead of absolute pixels to ensure they remain in the correct position despite a screen resize. Setting the transition duration to 0 is a lot faster than appending/detaching a class with a transition, which makes using the color picker a lot easier. Setting touch-action to none prevents the browser from handling events which should prevent the viewport from shifting when using the color picker